Create contact.html
Have a question or concern about [Your Discord Bot Name]? Want to provide feedback or suggest a new feature? We're here to listen! Please fill out the form above to send us a message. Be sure to include as much detail as possible, such as your Discord username and a description of your issue or suggestion. We'll do our best to respond promptly and provide helpful assistance.